webpage/backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
sqlparse==0.5.3 